content/base/test/test_bug654352-2.html

<!doctype html>
<html>
<!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=654352
-->
<head>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <title>Test for Bug 654352</title>
  <style>
    #a {
      padding: 10px;
      border: 8px solid black;
      width: 600px;
    }
  </style>
<script>
  function checkOffsetsFromPoint(aX, aY, aExpected) {
    var cp = document.mozCaretPositionFromPoint(aX, aY);
    ok(aExpected == cp.offset, 'expected offset at (' + aX + ', ' + aY + '): ' + aExpected + ', got: ' + cp.offset);
  }

  function doTesting() {
    // Tests at the beginning of the lines in the div
    checkOffsetsFromPoint(27, 35, 0);
    checkOffsetsFromPoint(25, 49, 1);

    // Test within the lines in the div
    checkOffsetsFromPoint(47, 57, 3);
    checkOffsetsFromPoint(115, 35, 11);

    /*
     *   These next two are questionably incorrect, as they both occur
     *   at the end of the two lines within the div. BUT, the
     *   new code matches the old code (focusOffset) for these,
     *   so it seems correct...
     */
    checkOffsetsFromPoint(143, 34, 1);
    checkOffsetsFromPoint(145, 51, 3);

    // Tests within the text area and input
    checkOffsetsFromPoint(45, 77, 2);
    checkOffsetsFromPoint(204, 106, 2);

    // Tests within the marquee
    checkOffsetsFromPoint(44, 148, 1);
    checkOffsetsFromPoint(103, 144, 7);

    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
</script>
</head>
<body onload="doTesting();">
<div id="a" contenteditable>abc, abc, abc<br>
abc, abc, abc<br>
<textarea>abc</textarea><input value="abc"><br><br>
<marquee>marquee</marquee>
</div>
</body>
</html>
